// src/api/generated/schema.ts
export interface components {
    schemas: {
        LoginTokenRequest: {
            token: string;
        };
        AuthSessionResponse: {
            authenticated: boolean;
            user?: components["schemas"]["AuthenticatedUser"];
        };
        AuthenticatedUser: {
            /** Format: int64 */
            id: number;
            username?: string | null;
            firstName?: string | null;
            lastName?: string | null;
            /** Format: uri */
            avatarUrl?: string | null;
        };
// ...
    };
    responses: never;
    parameters: {
        ReleaseId: number;
        TrackId: number;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}